cutSegments now processes indivisual files , moved batch processing to process function
